+++ STM32F1/libraries/RTClock/src/RTClock.h
@@ -41,7 +41,7 @@ class RTClock {
  public:
  	RTClock();
     RTClock(rtc_clk_src src ); 
-	RTClock(rtc_clk_src src, uint16 prescaler );
+	RTClock(rtc_clk_src src, uint32 prescaler );
 	//~RTClock(); //to implement
 	
 	void breakTime(time_t epoch_time, tm_t & tmm);
